# openapi/harness-api.yaml
openapi: 3.1.0
info:
  title: BILDR local API
  version: 1.0.0-draft
  description: >-
    Localhost-only API for the BILDR UI and harnessctl. This contract
    does not expose Codex App Server directly; harnessd projects it into durable
    repository/run/task/agent/evidence resources.
servers:
  - url: http://127.0.0.1:7310/api/v1
security:
  - LocalSession: []
...
paths:
...
  /repositories:
    get:
      tags: [Repositories]
      operationId: listRepositories
      responses:
        '200':
          description: Registered repositories.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Repository'}
    post:
      tags: [Repositories]
      operationId: registerRepository
      parameters:
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRepositoryRequest'}
      responses:
        '201':
          description: Repository registered.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Repository'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
  /repositories/{repositoryId}:
    get:
      tags: [Repositories]
      operationId: getRepository
      parameters:
        - {$ref: '#/components/parameters/RepositoryId'}
      responses:
        '200':
          description: Repository details and health.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Repository'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /repositories/{repositoryId}/inspect:
    post:
      tags: [Repositories]
      operationId: inspectRepository
      parameters:
        - {$ref: '#/components/parameters/RepositoryId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      responses:
        '202':
          description: Inspection started.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
...
  /runs:
    get:
      tags: [Runs]
      operationId: listRuns
      parameters:
        - name: repository_id
          in: query
          schema: {type: string}
        - name: state
          in: query
          schema: {type: string}
        - name: cursor
          in: query
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: Runs.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunPage'}
    post:
      tags: [Runs]
      operationId: createRun
      parameters:
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateRunRequest'}
      responses:
        '201':
          description: Prepared or preparing run.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /runs/{runId}:
    get:
      tags: [Runs]
      operationId: getRun
      parameters:
        - {$ref: '#/components/parameters/RunId'}
      responses:
        '200':
          description: Full run summary.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RunDetail'}
        '404': {$ref: '#/components/responses/NotFound'}
  /runs/{runId}/start-architecture:
    post:
      tags: [Runs]
      operationId: startArchitecture
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      responses:
        '202':
          description: Architect session queued.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
  /runs/{runId}/plan/approve:
    post:
      tags: [Runs]
      operationId: approvePlan
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [task_graph_digest]
              properties:
                task_graph_digest: {type: string}
                note: {type: string, maxLength: 4000}
                allow_budget_override:
                  type: boolean
                  default: false
                  description: Explicit local-user override when the remaining run ceiling is below the controller execution reserve.
      responses:
        '202':
          description: Scheduling enabled for approved graph.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /runs/{runId}/scheduler/pause:
    post:
      tags: [Runs]
      operationId: pauseRunScheduler
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      responses:
        '202':
          description: Scheduler paused.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
  /runs/{runId}/scheduler/resume:
    post:
      tags: [Runs]
      operationId: resumeRunScheduler
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                additional_token_budget:
                  type: integer
                  minimum: 0
                  maximum: 100000000
                  description: Optional governor allowance for the resumed window; zero uses the adaptive recommendation. Bounded child-thread headroom is reserved separately.
      responses:
        '202':
          description: Scheduler resumed and ready tasks dispatched. If the run budget was exhausted, the selected or adaptive governor allowance plus bounded child-thread headroom is added first.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
  /runs/{runId}/stop:
    post:
      tags: [Runs]
      operationId: stopRun
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [mode]
              properties:
                mode:
                  enum: [after_current_commands, interrupt_turns, cancel]
                preserve_all_worktrees: {type: boolean, default: true}
      responses:
        '202':
          description: Updated terminal or stopping run.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
  /runs/{runId}/archive:
    post:
      tags: [Runs]
      operationId: archiveRun
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      responses:
        '200':
          description: Terminal run archived; durable history and worktrees remain preserved.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Run'}
        '409': {$ref: '#/components/responses/Conflict'}
  /runs/{runId}/approve-integration:
    post:
      tags: [Runs]
      operationId: approveIntegration
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expected_head_sha]
              properties:
                expected_head_sha: {$ref: '#/components/schemas/GitSha'}
                note: {type: string, maxLength: 4000}
      responses:
        '202':
          description: Integration validation/final audit queued.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /runs/{runId}/publish-draft-pr:
    post:
      tags: [Runs]
      operationId: publishDraftPr
      parameters:
        - {$ref: '#/components/parameters/RunId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expected_head_sha, title]
              properties:
                expected_head_sha: {$ref: '#/components/schemas/GitSha'}
                title: {type: string, minLength: 1, maxLength: 240}
                body_appendix: {type: string, maxLength: 20000}
      responses:
        '202':
          description: Explicit external publication operation accepted.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /tasks/{taskId}/retry:
    post:
      tags: [Tasks]
      operationId: retryTask
      parameters:
        - {$ref: '#/components/parameters/TaskId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: {type: string, minLength: 1, maxLength: 4000}
                revised_objective: {type: string, maxLength: 4000}
                model_route:
                  enum: [same, escalate_terra]
                additional_token_budget: {type: integer, minimum: 0, maximum: 100000000}
      responses:
        '202':
          description: New immutable attempt created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
  /tasks/{taskId}/request-review:
    post:
      tags: [Tasks]
      operationId: requestTaskReview
      parameters:
        - {$ref: '#/components/parameters/TaskId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      responses:
        '202':
          description: Independent verifier queued.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
...
  /agents/{agentId}:
    get:
      tags: [Agents]
      operationId: getAgent
      parameters:
        - {$ref: '#/components/parameters/AgentId'}
      responses:
        '200':
          description: Agent session and current thread/turn.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AgentSummary'}
        '404': {$ref: '#/components/responses/NotFound'}
  /agents/{agentId}/steer:
    post:
      tags: [Agents]
      operationId: steerAgent
      parameters:
        - {$ref: '#/components/parameters/AgentId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [message]
              properties:
                message: {type: string, minLength: 1, maxLength: 12000}
                update_goal: {type: boolean, default: false}
      responses:
        '202':
          description: Steering accepted for delivery.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
  /agents/{agentId}/interrupt:
    post:
      tags: [Agents]
      operationId: interruptAgent
      parameters:
        - {$ref: '#/components/parameters/AgentId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      responses:
        '202':
          description: Interrupt requested.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OperationAccepted'}
...
  /approvals:
    get:
      tags: [Approvals]
      operationId: listApprovals
      parameters:
        - name: state
          in: query
          schema: {enum: [pending, resolved, expired]}
        - name: run_id
          in: query
          schema: {type: string}
      responses:
        '200':
          description: Approval requests.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Approval'}
  /approvals/{approvalId}/decision:
    post:
      tags: [Approvals]
      operationId: decideApproval
      parameters:
        - {$ref: '#/components/parameters/ApprovalId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision]
              properties:
                decision: {enum: [accept, decline, cancel, approve_once, deny]}
                note: {type: string, maxLength: 4000}
                expected_version: {type: integer, minimum: 1}
      responses:
        '202':
          description: Decision recorded and queued for runtime delivery.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '409': {$ref: '#/components/responses/Conflict'}
  /worktrees:
    get:
      tags: [Worktrees]
      operationId: listWorktrees
      parameters:
        - name: run_id
          in: query
          schema: {type: string}
        - name: state
          in: query
          schema: {type: string}
      responses:
        '200':
          description: Managed worktrees.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Worktree'}
  /worktrees/{worktreeId}/preserve:
    post:
      tags: [Worktrees]
      operationId: preserveWorktree
      parameters:
        - {$ref: '#/components/parameters/WorktreeId'}
        - {$ref: '#/components/parameters/CsrfToken'}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                reason: {type: string, maxLength: 4000}
      responses:
        '200':
          description: Worktree preservation updated.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Worktree'}
...
  /runs/{runId}/evidence:
    get:
      tags: [Evidence]
      operationId: listRunEvidence
      parameters:
        - {$ref: '#/components/parameters/RunId'}
      responses:
        '200':
          description: Evidence and claim matrix.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EvidenceSummary'}
...
components:
  securitySchemes:
    LocalSession:
      type: apiKey
      in: cookie
      name: harness_session
  parameters:
    CsrfToken:
      name: X-Harness-CSRF
      in: header
      required: true
      schema: {type: string, minLength: 16}
...
  responses:
...
    Conflict:
      description: State or optimistic-concurrency conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
...
    RegisterRepositoryRequest:
      type: object
      additionalProperties: false
      required: [profile_id, root_path]
      properties:
        profile_id: {type: string}
        root_path: {type: string}
        expected_origin: {type: string}
...
    Repository:
      type: object
      additionalProperties: false
      required: [id, profile_id, display_name, root_path, default_branch, health]
      properties:
        id: {type: string}
        profile_id: {type: string}
        display_name: {type: string}
        root_path: {type: string}
        origin_url: {type: string}
        default_branch: {type: string}
        primary_branch: {type: string}
        primary_head: {$ref: '#/components/schemas/GitSha'}
        primary_clean: {type: boolean}
        health: {enum: [ready, blocked, degraded]}
        blockers:
          type: array
          items: {type: string}
        managed_worktree_count: {type: integer}
        authority_digest: {type: string}
...